+++ src/store.rs
@@ -1,3 +1,4 @@
 pub mod bdd;
-mod common;
+pub mod common;
 pub mod dddmp;
+pub mod json;
