+++ src/store/bdd.rs
@@ -1,37 +1,9 @@
 use std::fs;
 
-use serde::{Deserialize, Serialize};
-
 use crate::{
-    core::{
-        bdd_manager::DDManager,
-        bdd_node::{DDNode, NodeID, VarID},
-    },
-    misc::hash_select::HashMap,
+    core::{bdd_manager::DDManager, bdd_node::NodeID},
+    store::common::{BddFile, Statistics},
 };
-
-type Statistics = Option<HashMap<NodeID, NodeStatistics>>;
-
-#[derive(Serialize, Deserialize)]
-struct BddFile {
-    statistics: Option<HashMap<String, NodeStatistics>>,
-    bdd: Bdd,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct NodeStatistics {
-    void: Option<bool>,
-    count: Option<usize>,
-}
-
-#[derive(Serialize, Deserialize)]
-struct Bdd {
-    order: Vec<usize>,
-    roots: Vec<NodeID>,
-    // node format:
-    // id_node id_var id_high id_low
-    nodes: String,
-}
 
 impl DDManager {
     /// Reads a (multi-rooted) BDD and corresponding statistics from a .bdd file.
@@ -48,70 +20,7 @@ impl DDManager {
         let bdd_file: BddFile =
             toml::from_str(&fs::read_to_string(filename).map_err(|e| e.to_string())?)
                 .map_err(|e| e.to_string())?;
-        let varorder: Vec<usize> = bdd_file.bdd.order;
-        let roots: Vec<NodeID> = bdd_file.bdd.roots;
-
-        let mut terminals: (Option<NodeID>, Option<NodeID>) = (None, None);
-        let nodes = bdd_file
-            .bdd
-            .nodes
-            .lines()
-            .filter(|line| !line.trim().is_empty())
-            .map(|line| {
-                let line: Vec<_> = line.split_whitespace().map(str::to_string).collect();
-                if line.len() != 4 {
-                    Err("Node list contains unexpected line!".to_string())
-                } else {
-                    let id = NodeID(line[0].parse::<usize>().map_err(|e| e.to_string())?);
-                    let varid = VarID(line[1].parse::<usize>().map_err(|e| e.to_string())?);
-                    let high = NodeID(line[2].parse::<usize>().map_err(|e| e.to_string())?);
-                    let low = NodeID(line[3].parse::<usize>().map_err(|e| e.to_string())?);
-                    if high == low {
-                        if high == NodeID(1) {
-                            terminals.0 = Some(id);
-                        } else if high == NodeID(0) {
-                            terminals.1 = Some(id);
-                        } else {
-                            return Err("Invalid terminal node!".to_string());
-                        }
-                    }
-                    Ok((id, (varid, high, low)))
-                }
-            })
-            .try_collect::<HashMap<NodeID, (VarID, NodeID, NodeID)>>()?;
-
-        // Check if root nodes are valid:
-        for r in roots.iter() {
-            if !nodes.contains_key(r) {
-                return Err("Root node not existant in BDD!".to_string());
-            }
-        }
-
-        // Check if terminal nodes exist:
-        let terminals = match terminals {
-            (Some(t_high), Some(t_low)) => Ok((t_high, t_low)),
-            _ => Err("Terminal nodes missing!".to_string()),
-        }?;
-
-        let (ddmanager, roots, id_translator) = DDManager::default()
-            .load_bdd_from_nodelist_with_translation(nodes, varorder, roots, terminals);
-
-        // Change node ids in statistic:
-        let statistics = bdd_file
-            .statistics
-            .map(|statistics| {
-                statistics
-                    .into_iter()
-                    .map(|(node_id, stats)| {
-                        let node_id =
-                            NodeID(node_id.parse::<usize>().map_err(|err| err.to_string())?);
-                        Ok::<_, String>((*id_translator.get(&node_id).unwrap(), stats))
-                    })
-                    .try_collect()
-            })
-            .transpose()?;
-
-        Ok((ddmanager, roots, statistics))
+        Self::load_from_bdd_file_object_with_statistics(bdd_file)
     }
 
     /// Reads a (multi-rooted) BDD from a .bdd file.
@@ -150,38 +59,8 @@ impl DDManager {
         roots: Vec<NodeID>,
         statistics: Statistics,
     ) -> Result<String, String> {
-        let order = self.var2level.clone();
-        let nodes = self
-            .nodes
-            .values()
-            .map(
-                |DDNode {
-                     id: NodeID(n),
-                     var: VarID(v),
-                     low: NodeID(l),
-                     high: NodeID(h),
-                     ..
-                 }| format!("{n} {v} {h} {l}"),
-            )
-            .fold("".to_string(), |mut s, line| {
-                s.push('\n');
-                s.push_str(&line);
-                s
-            });
-        let bdd_file = BddFile {
-            statistics: statistics.map(|statistics| {
-                statistics
-                    .into_iter()
-                    .map(|(node_id, stats)| (node_id.0.to_string(), stats))
-                    .collect()
-            }),
-            bdd: Bdd {
-                order,
-                roots,
-                nodes,
-            },
-        };
-        toml::to_string(&bdd_file).map_err(|e| e.to_string())
+        toml::to_string(&self.generate_bdd_file_object(roots, statistics))
+            .map_err(|e| e.to_string())
     }
 }
 
@@ -195,7 +74,7 @@ mod test {
             bdd_node::{DDNode, NodeID, VarID},
         },
         misc::hash_select::HashMap,
-        store::bdd::NodeStatistics,
+        store::common::NodeStatistics,
     };
 
     #[test]
